# information-tech/database/backend/db-service/src/main/java/com/university/db/dto/DatabaseDto.java
package com.university.db.dto;

import lombok.Data;
import org.springframework.hateoas.RepresentationModel;

@Data
public class DatabaseDto extends RepresentationModel<DatabaseDto> {

    private String id;
    private String name;

}