# information-tech/database/backend/db-service/src/main/java/com/university/db/dto/metadata/DatabaseMetadataDto.java
package com.university.db.dto.metadata;

import lombok.Data;
import org.springframework.hateoas.RepresentationModel;

import javax.validation.constraints.NotBlank;

@Data
public class DatabaseMetadataDto extends RepresentationModel<DatabaseMetadataDto> {

    @NotBlank
    private String name;

}