# information-tech/database/backend/db-service/src/main/resources/OpenAPI.yaml
openapi: 3.0.3
info:
  title: Database Application - 0.1.0
  description: |-
    Simple Database Engine
  version: 0.1.0
servers:
  - url: http://localhost:8080/
tags:
  - name: databases
    description: Operations with Databases
  - name: tables
    description: Operations with Tables
  - name: rows
    description: Operations with Rows
  - name: columns
    description: Operations with Columns
paths:
  /db-service/databases:
    get:
      tags:
        - databases
      summary: Find all databases
      operationId: findAllDatabases
      responses:
        '200':
          description: Successful found all databases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DatabaseDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: Database not found
        '405':
          description: Validation exception
    post:
      tags:
        - databases
      summary: Create a new database
      operationId: createDatabase
      responses:
        '201':
          description: Successfully created a database
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DatabaseDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: Database not found
        '405':
          description: Validation exception

  /db-service/databases/{id}:
    get:
      tags:
        - databases
      summary: Find database by id
      operationId: findDatabaseById
      parameters:
        - name: id
          in: path
          description: ID of the database to search for
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: Database not found
        '405':
          description: Validation exception
    put:
      tags:
        - databases
      summary: Edit an existing Database
      operationId: editDatabase
      parameters:
        - name: id
          in: path
          description: ID of the database to edit
          required: true
          schema:
            type: string
      requestBody:
        description: New database data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseMetadataDto'
        required: true
      responses:
        '200':
          description: Database was edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseDto'
        '400':
          description: Invalid body or id supplied
        '405':
          description: Validation exception
    delete:
      tags:
        - databases
      summary: Delete an existing database
      operationId: deleteDatabase
      parameters:
        - name: id
          in: path
          description: ID of the database to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Database was deleted
        '400':
          description: Invalid id supplied
        '405':
          description: Validation exception

  /db-service/databases/{id}/export:
    get:
      tags:
        - databases
      summary: Export an existing database
      operationId: exportDatabase
      parameters:
        - name: id
          in: path
          description: ID of the database to export
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Database was exported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportDatabaseDto'
        '400':
          description: Invalid id supplied
        '405':
          description: Validation exception

components:
  schemas:
    DatabaseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    DatabaseMetadataDto:
      type: object
      properties:
        name:
          type: string
    ExportDatabaseDto:
      type: object
      properties:
        name:
          type: string
        tables:
          $ref: '#/components/schemas/ExportTableDto'
    ExportTableDto:
      type: object
      properties:
        name:
          type: string
        columns:
          $ref: '#/components/schemas/ColumnMetadataDto'
        rows:
          $ref: '#/components/schemas/RowMetadataDto'
    ColumnMetadataDto:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - INTEGER
            - REAL
            - CHAR
            - STRING
            - TIME
            - TIMEINVL
    RowMetadataDto:
      type: object
      properties:
        fields:
          type: array
          items:
            type: string